split fle into its own mod
