+++ src/client/mod.rs
@@ -1,5 +1,7 @@
 pub mod auth;
 mod executor;
+#[cfg(feature = "fle")]
+mod fle;
 pub mod options;
 pub mod session;
 
@@ -9,7 +11,6 @@ use std::{
 };
 
 use derivative::Derivative;
-use mongocrypt::Crypt;
 
 #[cfg(test)]
 use crate::options::ServerAddress;
@@ -42,8 +43,6 @@ pub(crate) use executor::{HELLO_COMMAND_NAMES, REDACTED_COMMANDS};
 pub(crate) use session::{ClusterTime, SESSIONS_UNSUPPORTED_COMMANDS};
 
 use session::{ServerSession, ServerSessionPool};
-
-use self::options::AutoEncryptionOpts;
 
 const DEFAULT_SERVER_SELECTION_TIMEOUT: Duration = Duration::from_secs(30);
 
@@ -105,7 +104,7 @@ struct ClientInner {
     options: ClientOptions,
     session_pool: ServerSessionPool,
     #[cfg(feature = "fle")]
-    fle: Mutex<Option<FleState>>,
+    fle: Mutex<Option<fle::ClientState>>,
 }
 
 impl Client {
@@ -134,7 +133,7 @@ impl Client {
         let client = Self { inner };
         #[cfg(feature = "fle")]
         {
-            *client.inner.fle.lock().unwrap() = FleState::new(&client)?;
+            *client.inner.fle.lock().unwrap() = fle::ClientState::new(&client)?;
         }
         Ok(client)
     }
@@ -450,107 +449,5 @@ impl Client {
             .peek_latest()
             .description
             .clone()
-    }
-}
-
-#[cfg(feature = "fle")]
-#[derive(Derivative)]
-#[derivative(Debug)]
-struct FleState {
-    #[derivative(Debug = "ignore")]
-    crypt: Crypt,
-    mongocryptd_client: Option<Client>,
-    aux_clients: AuxClients,
-}
-
-#[cfg(feature = "fle")]
-#[derive(Debug)]
-struct AuxClients {
-    key_vault_client: Client,
-    metadata_client: Option<Client>,
-    internal_client: Option<Client>,
-}
-
-#[cfg(feature = "fle")]
-impl FleState {
-    fn new(client: &Client) -> Result<Option<Self>> {
-        let auto_enc_opts = match client.inner.options.auto_encryption_opts.as_ref() {
-            Some(o) => o,
-            None => return Ok(None),
-        };
-
-        let crypt = Self::make_crypt(auto_enc_opts)?;
-        let aux_clients = Self::make_aux_clients(client, auto_enc_opts)?;
-
-        Ok(Some(FleState {
-            crypt,
-            mongocryptd_client: None,
-            aux_clients,
-        }))
-    }
-
-    fn make_crypt(auto_enc_opts: &AutoEncryptionOpts) -> Result<Crypt> {
-        use std::path::Path;
-        let mut builder = Crypt::builder();
-        if Some(true) != auto_enc_opts.bypass_auto_encryption {
-            builder = builder.append_crypt_shared_lib_search_path(Path::new("$SYSTEM"))?;
-        }
-        if let Some(p) = auto_enc_opts
-            .extra_options
-            .as_ref()
-            .and_then(|o| o.get_str("cryptSharedLibPath").ok())
-        {
-            builder = builder.set_crypt_shared_lib_path_override(Path::new(p))?;
-        }
-        let crypt = builder.build()?;
-        if Some(true)
-            == auto_enc_opts
-                .extra_options
-                .as_ref()
-                .and_then(|o| o.get_bool("cryptSharedRequired").ok())
-        {
-            if crypt.shared_lib_version().is_none() {
-                return Err(crate::error::Error::invalid_argument(
-                    "cryptSharedRequired is set but crypt_shared is not available",
-                ));
-            }
-        }
-        Ok(crypt)
-    }
-
-    fn make_aux_clients(client: &Client, auto_enc_opts: &AutoEncryptionOpts) -> Result<AuxClients> {
-        let mut internal_client: Option<Client> = None;
-        let mut get_internal_client = || -> Result<Client> {
-            if let Some(c) = internal_client.clone() {
-                return Ok(c);
-            }
-            let mut internal_opts = client.inner.options.clone();
-            internal_opts.auto_encryption_opts = None;
-            internal_opts.min_pool_size = Some(0);
-            let c = Client::with_options(internal_opts)?;
-            internal_client = Some(c.clone());
-            Ok(c)
-        };
-
-        let key_vault_client = if let Some(c) = &auto_enc_opts.key_vault_client {
-            c.clone()
-        } else if Some(0) == client.inner.options.max_pool_size {
-            client.clone()
-        } else {
-            get_internal_client()?
-        };
-        let metadata_client = if Some(true) == auto_enc_opts.bypass_auto_encryption {
-            None
-        } else if Some(0) == client.inner.options.max_pool_size {
-            Some(client.clone())
-        } else {
-            Some(get_internal_client()?)
-        };
-
-        Ok(AuxClients {
-            key_vault_client,
-            metadata_client,
-            internal_client,
-        })
     }
 }
